requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==8.1.7
colorama==0.4.6
crispy-bootstrap4==2024.1
cryptography==43.0.3
Django==4.2.16
django-bootstrap-v5==1.0.11
django-bootstrap5==24.3
django-crispy-forms==2.3
django-filter==24.3
django-payments==3.0.1
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
flexcache==0.3
flexparser==0.4
idna==3.10
joblib==1.4.2
Markdown==3.7
mathparse==0.1.2
mercadopago==2.2.3
mysqlclient==2.2.4
nltk==3.9.1
paypalrestsdk==1.13.3
phonenumberslite==8.13.49
pillow==10.4.0
Pint==0.24.4
platformdirs==4.3.6
pycparser==2.22
pymongo==3.13.0
pyOpenSSL==24.2.1
python-dateutil==2.7.5
pytz==2024.2
PyYAML==3.13
razorpay==1.4.2
regex==2024.11.6
requests==2.32.3
six==1.16.0
soupsieve==2.6
SQLAlchemy==1.2.19
sqlparse==0.5.1
stripe==11.2.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
